Initialize project with a plan document
Prompt:

We are going to build a small web application for doing time zone conversions. The main UI
should be a single page where the user can select a timezone and a datetime range, as well
as one or more additional timezones. Then when that form is submitted, they will get back
that time range in the other selected timezones.

The backend will be a Python Flask app. The actual timezone conversion will be split out
into a separate file, for ease of testing.

The frontend will be simple HTML, using Bootstrap for styling. Native HTML validation should
be used whereever possible. Vanilla Javascript will be used only where absolutely necessary.
We will not use React or any other framework.

Start by creating a planning document describing the project and the work to be done.